+++ packages/cinterop/src/jvm/kotlin/io/realm/internal/interop/RealmInterop.kt
@@ -17,8 +17,13 @@
 package io.realm.internal.interop
 
 import io.realm.internal.interop.Constants.ENCRYPTION_KEY_LENGTH
-import io.realm.internal.interop.RealmInterop.cptr
+import io.realm.internal.interop.sync.AuthProvider
+import io.realm.internal.interop.sync.NetworkTransport
 import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.CoroutineStart
+import kotlinx.coroutines.launch
+import java.lang.reflect.Method
 
 // FIXME API-CLEANUP Rename io.realm.interop. to something with platform?
 //  https://github.com/realm/realm-kotlin/issues/56
@@ -38,7 +43,16 @@ actual object RealmInterop {
     // TODO API-CLEANUP Maybe pull library loading into separate method
     //  https://github.com/realm/realm-kotlin/issues/56
     init {
-        System.loadLibrary("realmc")
+        if (isAndroid()) {
+            System.loadLibrary("realmc")
+        } else {
+            // otherwise locate, using reflection, the dependency SoLoader and call load
+            // (calling SoLoader directly will create a circular dependency with `jvmMain`)
+            val classToLoad = Class.forName("io.realm.jvm.SoLoader")
+            val instance = classToLoad.newInstance()
+            val loadMethod: Method = classToLoad.getDeclaredMethod("load")
+            loadMethod.invoke(instance)
+        }
     }
 
     actual fun realm_get_version_id(realm: NativePointer): Long {
@@ -138,7 +152,13 @@ actual object RealmInterop {
     }
 
     actual fun realm_open(config: NativePointer, dispatcher: CoroutineDispatcher?): NativePointer {
-        val realmPtr = LongPointerWrapper(realmc.realm_open((config as LongPointerWrapper).ptr))
+        // create a custom Scheduler for JVM if a Coroutine Dispatcher is provided other wise pass null to use the generic one
+        val realmPtr = LongPointerWrapper(
+            realmc.open_realm_with_scheduler(
+                (config as LongPointerWrapper).ptr,
+                if (dispatcher != null) JVMScheduler(dispatcher) else null
+            )
+        )
         // Ensure that we can read version information, etc.
         realm_begin_read(realmPtr)
         return realmPtr
@@ -146,13 +166,6 @@ actual object RealmInterop {
 
     actual fun realm_freeze(liveRealm: NativePointer): NativePointer {
         return LongPointerWrapper(realmc.realm_freeze(liveRealm.cptr()))
-    }
-
-    actual fun realm_thaw(frozenRealm: NativePointer): NativePointer {
-        val realmPtr = LongPointerWrapper(realmc.realm_thaw(frozenRealm.cptr()))
-        // Ensure that we can read version information, etc.
-        realm_begin_read(realmPtr)
-        return realmPtr
     }
 
     actual fun realm_is_frozen(realm: NativePointer): Boolean {
@@ -216,12 +229,14 @@ actual object RealmInterop {
         return realmc.realm_object_is_valid(obj.cptr())
     }
 
-    actual fun realm_object_freeze(liveObject: NativePointer, frozenRealm: NativePointer): NativePointer {
-        return LongPointerWrapper(realmc.realm_object_freeze(liveObject.cptr(), frozenRealm.cptr()))
-    }
-
-    actual fun realm_object_thaw(frozenObject: NativePointer, liveRealm: NativePointer): NativePointer? {
-        return LongPointerWrapper(realmc.realm_object_thaw(frozenObject.cptr(), liveRealm.cptr()))
+    actual fun realm_object_resolve_in(obj: NativePointer, realm: NativePointer): NativePointer? {
+        val objectPointer = longArrayOf(0)
+        realmc.realm_object_resolve_in(obj.cptr(), realm.cptr(), objectPointer)
+        return if (objectPointer[0] != 0L) {
+            LongPointerWrapper(objectPointer[0])
+        } else {
+            null
+        }
     }
 
     actual fun realm_find_class(realm: NativePointer, name: String): ClassKey {
@@ -312,18 +327,21 @@ actual object RealmInterop {
         realmc.realm_list_erase(list.cptr(), index)
     }
 
-    actual fun realm_list_freeze(
-        liveList: NativePointer,
-        frozenRealm: NativePointer
-    ): NativePointer {
-        return LongPointerWrapper(realmc.realm_list_freeze(liveList.cptr(), frozenRealm.cptr()))
+    actual fun realm_list_resolve_in(
+        list: NativePointer,
+        realm: NativePointer
+    ): NativePointer? {
+        val listPointer = longArrayOf(0)
+        realmc.realm_list_resolve_in(list.cptr(), realm.cptr(), listPointer)
+        return if (listPointer[0] != 0L) {
+            LongPointerWrapper(listPointer[0])
+        } else {
+            null
+        }
     }
 
-    actual fun realm_list_thaw(
-        frozenList: NativePointer,
-        liveRealm: NativePointer
-    ): NativePointer {
-        return LongPointerWrapper(realmc.realm_list_thaw(frozenList.cptr(), liveRealm.cptr()))
+    actual fun realm_list_is_valid(list: NativePointer): Boolean {
+        return realmc.realm_list_is_valid(list.cptr())
     }
 
     // TODO OPTIMIZE Maybe move this to JNI to avoid multiple round trips for allocating and
@@ -435,8 +453,8 @@ actual object RealmInterop {
         realmc.realm_sync_client_config_set_base_file_path(syncClientConfig, basePath)
 
         // TODO add metadata mode to config
-        realmc.realm_sync_client_config_set_metadata_mode(syncClientConfig, realm_sync_client_metadata_mode_e.RLM_SYNC_CLIENT_METADATA_MODE_DISABLED)
-        return LongPointerWrapper(realmc.realm_app_new(appConfig.cptr(), syncClientConfig))
+        realmc.realm_sync_client_config_set_metadata_mode(syncClientConfig, realm_sync_client_metadata_mode_e.RLM_SYNC_CLIENT_METADATA_MODE_PLAINTEXT)
+        return LongPointerWrapper(realmc.realm_app_get(appConfig.cptr(), syncClientConfig))
     }
 
     actual fun realm_app_log_in_with_credentials(app: NativePointer, credentials: NativePointer, callback: CinteropCallback) {
@@ -448,12 +466,15 @@ actual object RealmInterop {
         )
     }
 
+    actual fun realm_network_transport_new(networkTransport: NetworkTransport): NativePointer {
+        return LongPointerWrapper(realmc.realm_network_transport_new(networkTransport))
+    }
     actual fun realm_app_config_new(
         appId: String,
-        networkTransportFactory: () -> Any,
+        networkTransport: NativePointer,
         baseUrl: String?
     ): NativePointer {
-        val config = realmc.new_app_config(appId, networkTransportFactory)
+        val config = realmc.realm_app_config_new(appId, networkTransport.cptr())
 
         baseUrl?.let { realmc.realm_app_config_set_base_url(config, it) }
 
@@ -472,8 +493,16 @@ actual object RealmInterop {
         realmc.realm_app_config_set_base_url(appConfig.cptr(), baseUrl)
     }
 
+    actual fun realm_app_credentials_new_anonymous(): NativePointer {
+        return LongPointerWrapper(realmc.realm_app_credentials_new_anonymous())
+    }
+
     actual fun realm_app_credentials_new_username_password(username: String, password: String): NativePointer {
         return LongPointerWrapper(realmc.realm_app_credentials_new_username_password(username, password))
+    }
+
+    actual fun realm_auth_credentials_get_provider(credentials: NativePointer): AuthProvider {
+        return AuthProvider.of(realmc.realm_auth_credentials_get_provider(credentials.cptr()))
     }
 
     actual fun realm_sync_config_new(user: NativePointer, partition: String): NativePointer {
@@ -524,12 +553,8 @@ actual object RealmInterop {
         return LongPointerWrapper(realmc.realm_query_find_all(query.cptr()))
     }
 
-    actual fun realm_results_freeze(liveResults: NativePointer, frozenRealm: NativePointer): NativePointer {
-        return LongPointerWrapper(realmc.realm_results_freeze(liveResults.cptr(), frozenRealm.cptr()))
-    }
-
-    actual fun realm_results_thaw(frozenResults: NativePointer, liveRealm: NativePointer): NativePointer {
-        return LongPointerWrapper(realmc.realm_results_thaw(frozenResults.cptr(), liveRealm.cptr()))
+    actual fun realm_results_resolve_in(results: NativePointer, realm: NativePointer): NativePointer {
+        return LongPointerWrapper(realmc.realm_results_resolve_in(results.cptr(), realm.cptr()))
     }
 
     actual fun realm_results_count(results: NativePointer): Long {
@@ -582,3 +607,24 @@ actual object RealmInterop {
         return Link(this.link.target_table, this.link.target)
     }
 }
+
+private class JVMScheduler(dispatcher: CoroutineDispatcher) {
+    val scope: CoroutineScope = CoroutineScope(dispatcher)
+
+    fun notifyCore(coreNotificationFunctionPointer: Long) {
+        val function: suspend CoroutineScope.() -> Unit = {
+            realmc.invoke_core_notify_callback(
+                coreNotificationFunctionPointer
+            )
+        }
+        scope.launch(
+            scope.coroutineContext,
+            CoroutineStart.DEFAULT,
+            function
+        )
+    }
+}
+
+// using https://developer.android.com/reference/java/lang/System#getProperties()
+private fun isAndroid(): Boolean =
+    System.getProperty("java.specification.vendor").contains("Android")