+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmConfigurationImpl.kt
@@ -51,9 +51,9 @@ public class RealmConfigurationImpl internal constructor(
 
     override val maxNumberOfActiveVersions: Long
 
-    override val notificationDispatcher: CoroutineDispatcher
+    val notificationDispatcher: CoroutineDispatcher
 
-    override val writeDispatcher: CoroutineDispatcher
+    val writeDispatcher: CoroutineDispatcher
 
     override val schemaVersion: Long